arm64: dts: qcom: sdm845-mtp: Make USB1 peripheral
The MTP is a "mobile reference device", as such the default operation is
to use fastboot to boot/flash software onto it and the common case is
thereby that we boot with a USB cable connected downstream from a PC or
a hub.  And without support for the PMI8998 charger block VBUS will not
be driven by the device.

Switch to peripheral until we can enable OTG.

Reviewed-by: Jeffrey Hugo <jeffrey.l.hugo@gmail.com>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Andy Gross <agross@kernel.org>

[ upstream commit: 9000a55bedb44f8fbb03bdf072dbd0657fc92a44 ]