dt-bindings: i2c: sun6i-p2wi: Fix the binding example
Even though the binding claims that the frequency can go up to 6MHz, the
common i2c binding sets a limit at 3MHz, which then triggers a warning.

Since the only SoC that uses that bus uses a frequency of 100kHz, and that
this bus hasn't been found in an SoC for something like 5 years, let's just
fix the example to have a frequency within the acceptable range for i2c.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>
Acked-by: Gregory CLEMENT <gregory.clement@bootlin.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>

[ upstream commit: 0038617132dea5c0516b6208a6d005ec9293ab94 ]