dt-bindings: net: Correct the documentation of KSZ9021 skew values
The documentation of skew values for the KSZ9021 PHY was misleading
because the driver implementation followed the erroneous information
given in the original KSZ9021 datasheet before it was corrected in
revision 1.2 (Feb 2014). It is probably too late to correct the driver
now because of the many existing device trees, so instead this just
corrects the documentation to explain that what you actually get is not
what you might think when looking at the device tree.

Signed-off-by: James Byrne <james.byrne@origamienergy.com>
Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>

[ upstream commit: 20b3f7d700130326db555d724e7507ce4606219a ]