ARM: dts: logicpd-torpedo: Disable Bluetooth Serial DMA
The default serial driver for omap2plus is the 8250_omap driver.
Unfortunately, this driver does not yet appear to have fully
functional DMA on OMAP3630/DM3730 which causes some timeouts and
frame errors.

This patch removes the DMA entry from the device tree which allow
the UART to operate without Bluetooth frame errors. If/when DMA
is working on OMAP3630, this should be reverted.

Signed-off-by: Adam Ford <aford173@gmail.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: 1b359d32f2b652032e056b79b20a7e1af7cf8ed2 ]