dt-bindings: Remove FIXME in yaml bindings
Some binding that were introduced early on got a comment to enable
additionalProperties, but we couldn't due to the generic properties being
reported as errors.

The way we're dealing with this now is to use the draft-08's
unevaluatedProperties (even though the tools doesn't do anything with it
yet).

Let's convert those old bindings to it.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: 454f5d9da1953e05c4df51b28c335daf49dd5425 ]