dt-bindings: clocks: Convert Allwinner legacy clocks to schemas
The Allwinner SoCs have a legacy set of bindings (and a framework to
support it in Linux) for their clock controllers.

Now that we have the DT validation in place, let's split into separate file
and convert the device tree bindings for those clocks to schemas, and mark
them all as deprecated.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: f95cad74acdb9de3b61a95ae8203c5e78b7d3615 ]