dt-bindings: resets: Convert Allwinner legacy resets to schemas
The Allwinner SoCs have a legacy set of bindings (and a framework to
support it in Linux) for their reset controllers.

Now that we have the DT validation in place, let's split into separate file
and convert the device tree bindings for those resets to schemas, and mark
them all as deprecated.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: 93adc6aef57f169c010071d732940b0f9f1fb5a7 ]