dt-bindings: Create DT bindings for SATA controllers
I need to create subnodes for drives connected to SATA
host controllers, and this needs to be supported
generally, so create a common YAML binding for
"sata" that will support subnodes with ports.

This has been designed as a subset of
ata/ahci-platform.txt with the bare essentials and
should be possible to extend or superset to cover the
common bindings.

Cc: Rob Herring <robh+dt@kernel.org>
Cc: devicetree@vger.kernel.org
Cc: Sergei Shtylyov <sergei.shtylyov@cogentembedded.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
[robh: fixup sata-port unit-address pattern]
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: 2b966a9d07375d966b738ecb3f8924e215e6273b ]